the fix
just type as int as there was already a check to ensure it will be an int

+++ astropy/convolution/utils.py
@@ -249,7 +249,7 @@ def discretize_oversample_1D(model, x_range, factor=10):
     # Evaluate model on oversampled grid
     x = np.linspace(x_range[0] - 0.5 * (1 - 1 / factor),
                     x_range[1] - 0.5 * (1 + 1 / factor),
-                    num = (x_range[1] - x_range[0]) * factor)
+                    num=int((x_range[1] - x_range[0]) * factor))
 
     values = model(x)
 
@@ -265,10 +265,10 @@ def discretize_oversample_2D(model, x_range, y_range, factor=10):
     # Evaluate model on oversampled grid
     x = np.linspace(x_range[0] - 0.5 * (1 - 1 / factor),
                     x_range[1] - 0.5 * (1 + 1 / factor),
-                    num = (x_range[1] - x_range[0]) * factor)
+                    num=int((x_range[1] - x_range[0]) * factor))
     y = np.linspace(y_range[0] - 0.5 * (1 - 1 / factor),
                     y_range[1] - 0.5 * (1 + 1 / factor),
-                    num = (y_range[1] - y_range[0]) * factor)
+                    num=int((y_range[1] - y_range[0]) * factor))
 
     x_grid, y_grid = np.meshgrid(x, y)
     values = model(x_grid, y_grid)
